Don't enforce inside returns with non-nil errors
If a function has a struct return type alongside an error,
it's pretty standard to do this:

    if err != nil {
        return MyStruct{}, fmt.Errorf("do thing: %w", err)
    }

There's no reason to enforce required fields on the struct
because the contract is that the struct value is to be ignored
if the error is non-nil.

This change adds support to detect these cases and ignore them.
It is a bit messier than "ignore if part of return with non-nil error"
because we still want to enforce required fields if the non-nil error
itself is the thing with required fields, e.g.

    return nil, &MyError{...} // and MyError has required fields

Resolves #37