Fix function type resolution with nested funcs
Fixes #57

isReturnedWithNonNilError walks up from the return statement to find
the matching function type, but it doesn't `break` when it finds the
closest function type, which can lead to it using the wrong signature.

Move the function matching into a separate function to avoid a label
on the loop so we can `break` correctly.